python-ai/requirements.txt
flask
flask-cors
anthropic
pandas
scikit-learn
python-dotenv
